Renamed extractor to extract-examples
