+++ src/extract-examples.js
